Cleanup types in service for cost-tracking and cost calc via builder
